Fix circular dependency between printer and transforms in plugin workspace
