--- src/plugin/src/printer/doc-comment/doc-comment-options.ts
@@ -1,20 +1,0 @@
-import { Core } from "@gml-modules/core";
-
-import { DEFAULT_PRINT_WIDTH } from "../../constants.js";
-
-type DocCommentPrinterOptions = Record<string, unknown> & {
-    printWidth?: number;
-};
-
-type ResolvedDocCommentPrinterOptions = DocCommentPrinterOptions & {
-    printWidth: number;
-};
-
-export function resolveDocCommentPrinterOptions(options?: DocCommentPrinterOptions): ResolvedDocCommentPrinterOptions {
-    const printWidth = Core.coercePositiveIntegerOption(options?.printWidth, DEFAULT_PRINT_WIDTH);
-
-    return {
-        ...options,
-        printWidth
-    };
-}
