+++ src/plugin/src/printer/doc-comment/function-docs.ts
@@ -1,10 +1,9 @@
 import { Core, type MutableDocCommentLines } from "@gml-modules/core";
 
 import { formatDocLikeLineComment } from "../../comments/index.js";
-import { removeFunctionDocCommentLines } from "../../doc-comment/index.js";
+import { removeFunctionDocCommentLines, resolveDocCommentPrinterOptions } from "../../doc-comment/index.js";
 import { DescriptionUtils, NormalizationUtils } from "../../transforms/doc-comment/index.js";
 import { safeGetParentNode } from "../path-utils.js";
-import { resolveDocCommentPrinterOptions } from "./doc-comment-options.js";
 
 const STRING_TYPE = "string";
 const BLANK_LINE_PATTERN = /(?:\r\n|\r|\n|\u2028|\u2029)\s*(?:\r\n|\r|\n|\u2028|\u2029)/;
